Move DTO validation logic to DTOs module
